Added builders. Person Test class implemented and working. Modifications in Person Controller class methods

+++ LAB07/myrestproject/src/main/java/myrestproject/controller/PersonRestController.java
@@ -10,6 +10,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -30,6 +33,7 @@ public class PersonRestController {
 		return new ResponseEntity<List<Person>>(personService.getAllPersons(),HttpStatus.OK);		
 	}
 	
+
 	@GetMapping("/person/{id}")
 	public ResponseEntity<Person> getPerson(@PathVariable int id) {
 		return new ResponseEntity<Person>(personService.findById(id),HttpStatus.OK);
@@ -39,13 +43,22 @@ public class PersonRestController {
 	public ResponseEntity<Person> addPerson(@RequestParam String name, @RequestParam int yob) {
 		Person p = new Person(name, yob);
 		personService.addPerson(p);
-		return new ResponseEntity<Person>(p,HttpStatus.OK);
+		return new ResponseEntity<Person>(p,HttpStatus.CREATED);
 	}
 	
-	@GetMapping("/deleteperson")
-	public ResponseEntity<Person> deletePerson(@RequestParam int id) {
-		personService.deletePerson(id);
-		return new ResponseEntity<Person>(HttpStatus.OK);
+	@RequestMapping(value="/addperson",method=RequestMethod.POST)
+	public ResponseEntity<Person> postPerson(@RequestBody Person person) {
+		personService.addPerson(person);
+		return new ResponseEntity<Person>(person,HttpStatus.CREATED);
+	}
+	
+	@RequestMapping(value="/person/{id}", method=RequestMethod.DELETE)
+	public ResponseEntity<Person> deletePerson(@PathVariable int id) {
+		if(personService.findById(id) != null) {
+			personService.deletePerson(id);
+			return new ResponseEntity<Person>(HttpStatus.OK);
+		}
+		else return new ResponseEntity<Person>(HttpStatus.NO_CONTENT);
 	}
